backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
pandas==2.2.2
python-multipart==0.0.9
pydantic==2.7.1
google-generativeai==0.7.2
requests==2.31.0